+++ src/components/organisms/SignIn/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Link } from "react-router-dom";
+import { useHistory } from "react-router";
 import styled from "styled-components";
 import { removeError, signIn } from "../../../modules/session";
 import Box from "../../atoms/Box";
@@ -8,6 +9,7 @@ import SpanError from "../../atoms/SpanError";
 import SpanLoading from "../../atoms/SpanLoading";
 import Logo from "../../atoms/Logo";
 import Form from "../../molecules/Form";
+import Span from "../../atoms/Span";
 
 const StyledDiv = styled.div`
   width: ${({ theme }) => theme.widths.loginBox};
@@ -18,58 +20,32 @@ const StyledSpan = styled.span`
   margin: 10px 0;
 `;
 
-const LinkSpan = styled.span`
-  color: ${({ theme }) => theme.colors.blue};
-`;
-
-const SignIn = () => {
-  const dispatch = useDispatch();
+const SignIn = ({ signInProps, signUpProps }) => {
+  const { onSubmit, onClickLink, FormProps } = signInProps;
   const { loading, error } = useSelector((state) => state.session);
 
-  const onSubmit = (e) => {
-    e.preventDefault(); // 이걸 안하면 주소에 query(맞나?)가 붙어버린다!
-
-    dispatch(
-      signIn({
-        username: e.target.username.value,
-        password: e.target.password.value,
-      })
-    );
-  };
-
-  const onClickLink = () => {
-    dispatch(removeError());
-  };
-
-  const FormProps = {
-    inputInfo: {
-      username: "사용자 이름",
-      password: "비밀번호",
-    },
-    btnName: "로그인",
-  };
-
-  return (
+  return loading ? (
+    <SpanLoading />
+  ) : (
     <StyledDiv onSubmit={onSubmit}>
-      {loading ? (
-        <SpanLoading />
-      ) : (
-        <>
-          <Box>
-            <Logo />
-            <Form {...FormProps} />
-            {error ? <SpanError>{error}</SpanError> : null}
-          </Box>
-          <Box>
-            <StyledSpan>
-              계정이 없으신가요?{" "}
-              <Link to="/accounts/signup" onClick={onClickLink}>
-                <LinkSpan>가입하기</LinkSpan>
-              </Link>
-            </StyledSpan>
-          </Box>
-        </>
-      )}
+      <Box>
+        <Logo />
+        <Form {...FormProps} />
+        {error ? <SpanError>{error}</SpanError> : null}
+      </Box>
+      <Box>
+        <StyledSpan>
+          계정이 없으신가요?{" "}
+          <Link
+            to={{ pathname: "/accounts/signup", state: { ...signUpProps } }}
+            onClick={onClickLink}
+          >
+            <Span fontSize="14px" color="blue">
+              가입하기
+            </Span>
+          </Link>
+        </StyledSpan>
+      </Box>
     </StyledDiv>
   );
 };
